+++ showcase/shell-dashboard/src/components/dashboard-page.tsx
@@ -1,0 +1,292 @@
+"use client";
+// Feature matrix: composable overlay-driven 2-tab layout.
+// Matrix tab: overlay toggles control which visual layers render.
+// Ops tab: probe status grid (unchanged from legacy).
+//
+// This is the client half of the dashboard route. The route file
+// (`src/app/page.tsx`) is a SERVER component that reads `SHELL_URL` at
+// request time via `getRuntimeConfig()` and threads it down here as the
+// `shellUrl` prop. Threading the real host server-side (rather than
+// reading the client config, which returns the
+// `https://ssr-placeholder.invalid/` sentinel during SSR) means Demo /
+// Code anchors are built with the REAL host in the initial HTML — the
+// sentinel never leaks into the rendered DOM, and links work pre-hydration
+// (crawlers, no-JS). See FeatureGrid's `shellUrl` prop.
+import { useCallback, useEffect, useMemo, useState } from "react";
+import { FeatureGrid } from "@/components/feature-grid";
+import type { CellContext } from "@/components/feature-grid";
+import { StatusTab } from "@/components/status-tab";
+import { useLiveStatus } from "@/hooks/useLiveStatus";
+import { useProbes, useTriggerProbe } from "@/hooks/use-probes";
+import { mergeRowsToMap } from "@/lib/live-status";
+import { useOverlays } from "@/hooks/useOverlays";
+import { OverlayToggleBar } from "@/components/overlay-toggle-bar";
+import { UnifiedCell } from "@/components/unified-cell";
+import { buildCellModel } from "@/lib/cell-model";
+import {
+  computeHealthStats,
+  computeParityStats,
+  computeDepthDistribution,
+  computeD6Stats,
+} from "@/lib/page-stats";
+import { AdaptiveStatsBar } from "@/components/adaptive-stats-bar";
+import { AdaptiveLegend } from "@/components/adaptive-legend";
+import { ThemeToggle } from "@/components/theme-toggle";
+import { BaselineTab } from "@/components/baseline-tab";
+import { DiscoveryAuthBanner } from "@/components/discovery-auth-banner";
+import { getDocsStatus } from "@/lib/docs-status";
+import catalog from "@/data/catalog.json";
+import type { CatalogData } from "@/data/catalog-types";
+
+// Single cast: catalog.json is generated by `pretest`/`predev`/`prebuild` from
+// generate-registry.ts and conforms to CatalogData by construction.
+const catalogData: CatalogData = catalog as CatalogData;
+
+export interface DashboardPageProps {
+  /**
+   * Shell host resolved server-side (request-time `SHELL_URL`) by the route
+   * server component and threaded down to FeatureGrid so Demo / Code anchors
+   * carry the REAL host during SSR — no `ssr-placeholder.invalid` sentinel.
+   */
+  shellUrl: string;
+}
+
+export function DashboardPage({ shellUrl }: DashboardPageProps) {
+  // Single subscription for ALL dimensions — no filter. Avoids the 7-way
+  // SSE race that caused PB subscribe 400s when 7 hooks competed for the
+  // single PB SDK realtime connection during hydration.
+  const allStatus = useLiveStatus();
+
+  const liveStatus = useMemo(
+    () => mergeRowsToMap(allStatus.rows),
+    [allStatus.rows],
+  );
+
+  const connection = allStatus.status;
+
+  // Wall-clock tick: forces staleness re-evaluation even when no SSE delta
+  // arrives. Without this, a wedged pipeline (no new rows) would never
+  // re-sample time, so a frozen-green cell could never downgrade to stale.
+  // One interval, 60s cadence (well under the staleness window), cleared on
+  // unmount. Bumping `tick` only invalidates the `now` memo below, so the
+  // re-render is cheap — the same recompute that already runs on every SSE
+  // delta.
+  const [tick, setTick] = useState(0);
+  useEffect(() => {
+    const id = setInterval(() => setTick((t) => t + 1), 60_000);
+    return () => clearInterval(id);
+  }, []);
+
+  // Single reference time shared by EVERY staleness-deriving call below
+  // (healthStats, depthDistribution, d6Stats, renderCell). Re-sampled when
+  // live-status changes OR the 60s `tick` fires, so chips and stats agree on
+  // which green rows are stale instead of each defaulting to its own
+  // `Date.now()` that may have crossed a window boundary milliseconds later.
+  // Mirrors the single-`now` discipline in cell-matrix.tsx.
+  const now = useMemo(() => Date.now(), [liveStatus, tick]);
+
+  // R2-D.1: real probe wiring. `useProbes` polls the ops API every 10s and
+  // feeds the schedule grid; `useTriggerProbe` POSTs to /trigger with the
+  // operator token. If the token is unset, the trigger callback throws
+  // a clear "token required" error from useTriggerProbe — fail-loud so
+  // operators see the issue rather than a silent no-op.
+  const probesQuery = useProbes();
+  const probeEntries = useMemo(
+    () => probesQuery.data?.probes ?? [],
+    [probesQuery.data],
+  );
+  const triggerToken = process.env.NEXT_PUBLIC_OPS_TRIGGER_TOKEN;
+  const { trigger } = useTriggerProbe({ token: triggerToken });
+  const handleTrigger = useCallback(
+    async (probeId: string, slugs?: string[]): Promise<void> => {
+      await trigger(probeId, slugs);
+    },
+    [trigger],
+  );
+
+  // Overlay state management (handles URL hash + localStorage persistence)
+  const {
+    overlays,
+    activeTab,
+    toggle,
+    applyPreset,
+    setTab,
+    activePreset,
+    selectedProbeId,
+    selectProbe,
+  } = useOverlays();
+
+  // Compute health stats using buildCellModel — single source of truth for
+  // chip color, so the stats bar matches what's visually displayed in the
+  // table. A gray chip (no data yet) is tracked as `noData`, NOT folded into
+  // green, so the bar agrees with the matrix it summarizes.
+  const healthStats = useMemo(
+    () => computeHealthStats(catalogData.cells, liveStatus, now),
+    [liveStatus, now],
+  );
+
+  // Compute parity tier counts (validates tier before indexing — fail-loud
+  // on unknown tiers rather than producing NaN).
+  const parityStats = useMemo(() => computeParityStats(catalogData.cells), []);
+
+  // Compute docs reachability stats from unique feature IDs
+  const docsStats = useMemo(() => {
+    let ok = 0;
+    let missing = 0;
+    let notFound = 0;
+    let error = 0;
+    const seen = new Set<string>();
+    for (const cell of catalogData.cells) {
+      if (cell.feature !== null && !seen.has(cell.feature)) {
+        seen.add(cell.feature);
+        const status = getDocsStatus(cell.feature);
+        if (status.og === "ok") ok++;
+        else if (status.og === "missing") missing++;
+        else if (status.og === "notfound") notFound++;
+        else if (status.og === "error") error++;
+      }
+    }
+    return { ok, missing, notFound, error };
+  }, []);
+
+  // Compute depth distribution across all wired cells — re-derives whenever
+  // live-status changes since depth is a function of probe states. Uses an
+  // exhaustive achievedDepth→key map so D6-achieved cells land in the `d6`
+  // bucket (the old string cast dropped them).
+  const depthDistribution = useMemo(
+    () => computeDepthDistribution(catalogData.cells, liveStatus, now),
+    [liveStatus, now],
+  );
+
+  // Compute D6 (parity-vs-reference) rollup counts across wired cells —
+  // degraded/stale D6 surfaces distinctly instead of folding into gray.
+  const d6Stats = useMemo(
+    () => computeD6Stats(catalogData.cells, liveStatus, now),
+    [liveStatus, now],
+  );
+
+  // renderCell callback wrapping UnifiedCell + buildCellModel.
+  // buildCellModel is the single source of truth for cell state — it handles
+  // not-supported features, ceiling detection, and chip color derivation.
+  const renderCell = useCallback(
+    (ctx: CellContext) => {
+      const isSupported = !(
+        ctx.integration.not_supported_features?.includes(ctx.feature.id) ??
+        false
+      );
+      const model = buildCellModel(
+        ctx.liveStatus,
+        {
+          slug: ctx.integration.slug,
+          featureId: ctx.feature.id,
+          isSupported,
+          isWired: true, // renderCell only called when demo exists
+        },
+        now,
+      );
+      return <UnifiedCell ctx={ctx} model={model} overlays={overlays} />;
+    },
+    [overlays, now],
+  );
+
+  return (
+    <div data-testid="tab-shell" className="h-dvh flex flex-col">
+      {/* Tab bar — pinned above scroll area */}
+      <div className="flex-shrink-0 flex items-center gap-0 border-b border-[var(--border)] px-8">
+        <button
+          type="button"
+          data-testid="tab-matrix"
+          onClick={() => setTab("matrix")}
+          className={`px-4 py-2.5 text-sm font-medium transition-colors cursor-pointer ${
+            activeTab === "matrix"
+              ? "text-[var(--accent)] border-b-2 border-[var(--accent)] -mb-px"
+              : "text-[var(--text-muted)] hover:text-[var(--text-secondary)]"
+          }`}
+        >
+          Coverage
+        </button>
+        <button
+          type="button"
+          data-testid="tab-baseline"
+          onClick={() => setTab("baseline")}
+          className={`px-4 py-2.5 text-sm font-medium transition-colors cursor-pointer ${
+            activeTab === "baseline"
+              ? "text-[var(--accent)] border-b-2 border-[var(--accent)] -mb-px"
+              : "text-[var(--text-muted)] hover:text-[var(--text-secondary)]"
+          }`}
+        >
+          Baseline
+        </button>
+        <button
+          type="button"
+          data-testid="tab-ops"
+          onClick={() => setTab("ops")}
+          className={`px-4 py-2.5 text-sm font-medium transition-colors cursor-pointer ${
+            activeTab === "ops"
+              ? "text-[var(--accent)] border-b-2 border-[var(--accent)] -mb-px"
+              : "text-[var(--text-muted)] hover:text-[var(--text-secondary)]"
+          }`}
+        >
+          Ops
+        </button>
+        <div className="ml-auto flex items-center pr-1">
+          <ThemeToggle />
+        </div>
+      </div>
+
+      <DiscoveryAuthBanner rows={allStatus.rows} />
+
+      {activeTab === "matrix" && (
+        <>
+          {/* Overlay toggle — pinned below tab bar */}
+          <div className="flex-shrink-0 px-8 py-3 bg-[var(--bg-surface)] border-b border-[var(--border)]">
+            <OverlayToggleBar
+              overlays={overlays}
+              onToggle={toggle}
+              onApplyPreset={applyPreset}
+              activePreset={activePreset}
+            />
+          </div>
+          {/* Single scroll area — stats bar and title scroll away, table headers stick */}
+          <div className="flex-1 min-h-0 overflow-auto pb-12">
+            <div className="px-8 py-3 border-b border-[var(--border)]">
+              <AdaptiveStatsBar
+                overlays={overlays}
+                catalog={catalogData}
+                healthStats={healthStats}
+                parityStats={parityStats}
+                docsStats={docsStats}
+                depthDistribution={depthDistribution}
+                d6Stats={d6Stats}
+              />
+            </div>
+            <FeatureGrid
+              title="Feature Matrix"
+              renderCell={renderCell}
+              minColWidth={180}
+              liveStatus={liveStatus}
+              connection={connection}
+              overlays={overlays}
+              catalog={catalogData}
+              shellUrl={shellUrl}
+            />
+            <AdaptiveLegend overlays={overlays} />
+          </div>
+        </>
+      )}
+
+      {activeTab === "baseline" && <BaselineTab />}
+
+      {activeTab === "ops" && (
+        <div className="flex-1 min-h-0 overflow-auto">
+          <StatusTab
+            entries={probeEntries}
+            onTrigger={handleTrigger}
+            selectedProbeId={selectedProbeId}
+            onSelectProbe={selectProbe}
+          />
+        </div>
+      )}
+    </div>
+  );
+}
